crud completed

+++ src/app/components/clients/clients.component.ts
@@ -1,9 +1,12 @@
+import { ConfirmDialogComponent } from './../confirm-dialog/confirm-dialog.component';
+import { ClientDialogComponent } from './../client-dialog/client-dialog.component';
 import { ClientService } from './../../services/client.service';
 import { Client } from './../../models/client.model';
 import { ChangeDetectorRef, Component, OnInit, ViewChild } from '@angular/core';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-clients',
@@ -20,7 +23,8 @@ export class ClientsComponent implements OnInit {
 
   constructor(
     private _client: ClientService,
-    private _cdr: ChangeDetectorRef
+    private _cdr: ChangeDetectorRef,
+    public dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -33,16 +37,31 @@ export class ClientsComponent implements OnInit {
       this.dataSource = new MatTableDataSource<Client>(this.clients);
       this.dataSource.paginator = this.paginator;
       this.dataSource.sort = this.sort;
-      console.log(this.clients, this.dataSource, this.paginator);
+      // console.log(this.clients, this.dataSource, this.paginator);
     });
   }
 
   deleteClient(id: string) {
-    console.log(id);
-    this._client.deleteClient(id).subscribe((res) => {
-      console.log(res);
-      this.getClients();
-      this._cdr.detectChanges();
+    // console.log(id);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent);
+
+    dialogRef.afterClosed().subscribe((result) => {
+      console.log(`Dialog result: ${result}`);
+      if (result) {
+        this._client.deleteClient(id).subscribe((res) => {
+          console.log(res);
+          this.getClients();
+          this._cdr.detectChanges();
+        });
+      }
+    });
+  }
+
+  onMoreInfo(id: string) {
+    const client: Client = this.clients.find((client) => client._id === id)!;
+    // console.log(id, client, this.clients);
+    const dialogRef = this.dialog.open(ClientDialogComponent, {
+      data: client,
     });
   }
 }
